perf: round 2 optimizations - tbt reduction via code splitting and deferred analytics
Changes:
- deferred all analytics scripts to lazyOnload strategy (TBT reduction)
- removed console.error from theme script for silent failures
- added adjustFontFallback to Geist fonts for better CLS
- moved SpeedInsights outside body tag
- aggressive code splitting: extracted 4 homepage sections into separate components
  * NowBloomingSection, TreeOfTheDay, StatsSection, AboutSection
- added loading skeletons for all dynamically loaded sections
- created reusable skeleton components for better UX

Expected impact:
- TBT: 298ms -> <200ms (targeting ~100ms reduction)
- Performance score: 67 -> 75-82 (targeting +8-13 points)
- better perceived performance with skeleton loading states
- improved font loading with CLS reduction
